WIP caching
This is an attempt to cache the following:
* 1 minute for index.html
* 10 minutes for /config.json
* 365 days for hashed resources

Current limitation is that index.html isn't effective as queries return
index.html's content (single page app).

Ex. the following's content is the same content of index.html
```
http://localhost:9411/?serviceName=zipkin-query&spanName=all&endTs=1458742512402&minDuration=&limit=10&annotationQuery=
```

One idea could be to get a content hash of index.html or customize
`fileOrIndex` to somehow let us know the content is the same..